RoaringArray64: fix architecture for input values

+++ roaring64/roaringarray64_test.go
@@ -39,11 +39,10 @@ func TestCopies(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			r1 := 1 << 32
+			r1 := uint64(1) << 32
 			r2 := uint64(2) << 32
 			r3 := uint64(3) << 32
 			r4 := uint64(4) << 32
-			// key := highbits(uint64(r4))
 
 			bitmap1 := New()
 			bitmap2 := New()
@@ -91,10 +90,10 @@ func TestCopies(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			r1 := 1 << 32
-			r2 := 2 << 32
-			r3 := 3 << 32
-			r4 := 4 << 32
+			r1 := uint64(1) << 32
+			r2 := uint64(2) << 32
+			r3 := uint64(3) << 32
+			r4 := uint64(4) << 32
 			key := highbits(uint64(r4))
 
 			bitmap1 := New()
